+++ CHANGELOG.md
@@ -18,6 +18,7 @@
     - root/panel/modal/tab: support setting `scroll` to `hidden` (hides scrollbars and allows scrolling)
     - add new `scrollbar` widget that connects to a scrolling container widget and acts as it scrollbar
     - input: add `noCanvas` option to avoid blurry text under some circumstances
+    - xy: add `clipX` and `clipY` proprties (allows multixy to set per-pads anges limits)
 
 - editor
     - selecting a widget automatically shows it in the widget tree if it's not minimized